streamlit/requirements.txt
streamlit
pandas
requsets
urllib
PILLOW
openai
youtube_transcript_api
couchbase
